No ilq game timing in pythong
It is way too slow and the python version does not allow for regularization and
has a bug in the solver (does not consider linear state cost terms in quadratization)
Thus, no timing for this.
The single iterate comparison already shows, that python is way too slow for this.

For the record: a single iterate of the python implementation takes about 4.254564s